Light Observable UI Alpha version
Light Observable UI is a lightweight MVVM library, written in
TypeScript, that provides features to ease the development of modern
single page applications.

Main features:
* Page - A page takes the entire screen.
* Content view - A content view is the building block of a page.
* Templating - Write JSX in .tsx files.
* Binding - Bi-directional binding between HTML views and view models.
* MessagingCenter - Provides decoupled event handling for communication
  	between components that do not know about each other. Pub / Sub.
* HttpClient - Built-in HTTP client.
* Validation - Easy to use validation of input data.

Light Observable UI is still in the very early stages of development.
It is far from production ready. Every program begins with a single
commit.